[#81] planned relations as dotted line
