+++ src/main/java/de/bonndan/nivio/output/map/ItemMapItem.java
@@ -1,10 +1,18 @@
 package de.bonndan.nivio.output.map;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import de.bonndan.nivio.model.*;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import de.bonndan.nivio.model.Item;
+import de.bonndan.nivio.model.LandscapeItem;
+import de.bonndan.nivio.model.RelationItem;
+import de.bonndan.nivio.model.Status;
 import de.bonndan.nivio.output.Color;
 import org.springframework.util.StringUtils;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,8 +23,7 @@ import java.util.stream.Collectors;
  * The x,y coordinates are derived from the rendered representation.
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public
-class ItemMapItem extends MapItem {
+public class ItemMapItem extends MapItem {
 
     public long x;
     public long y;
@@ -46,16 +53,27 @@ class ItemMapItem extends MapItem {
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     public static class Relation {
-        public final String source;
-        public final String target;
+
+        @JsonSerialize(using = ItemSerializer.class)
+        public final Item source;
+        @JsonSerialize(using = ItemSerializer.class)
+        public final Item target;
+
         public final String type;
         public final String format;
 
         public Relation(RelationItem<Item> rel) {
-            this.source = rel.getSource().getFullyQualifiedIdentifier().toString();
-            this.target = rel.getTarget().getFullyQualifiedIdentifier().toString();
+            this.source = rel.getSource();
+            this.target = rel.getTarget();
             this.type = rel.getType() != null ? rel.getType().name() : null;
             this.format = rel.getFormat();
+        }
+
+        private static class ItemSerializer extends JsonSerializer<Item> {
+            @Override
+            public void serialize(Item value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
+                gen.writeString(value.getFullyQualifiedIdentifier().toString());
+            }
         }
     }
 }
