+++ src/main/java/de/bonndan/nivio/output/map/svg/NPattern.java
@@ -1,4 +1,4 @@
-package de.bonndan.nivio.output.map;
+package de.bonndan.nivio.output.map.svg;
 
 import j2html.tags.DomContent;
