+++ src/main/java/de/bonndan/nivio/output/map/svg/SVGItemLabel.java
@@ -1,5 +1,6 @@
-package de.bonndan.nivio.output.map;
+package de.bonndan.nivio.output.map.svg;
 
+import de.bonndan.nivio.output.map.ItemMapItem;
 import j2html.tags.ContainerTag;
 import j2html.tags.DomContent;
 import org.springframework.util.StringUtils;
