+++ src/main/java/de/bonndan/nivio/output/map/svg/SVGRelation.java
@@ -1,5 +1,7 @@
-package de.bonndan.nivio.output.map;
+package de.bonndan.nivio.output.map.svg;
 
+import de.bonndan.nivio.model.Lifecycle;
+import de.bonndan.nivio.output.map.ItemMapItem;
 import j2html.tags.ContainerTag;
 import j2html.tags.DomContent;
 
@@ -8,7 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import static de.bonndan.nivio.output.map.SvgTagCreator.g;
+import static de.bonndan.nivio.output.map.svg.SvgTagCreator.g;
 
 class SVGRelation extends Component {
 
@@ -32,15 +34,16 @@ class SVGRelation extends Component {
         bezierPath.parsePathString(stringPath);
 
         if ("PROVIDER".equals(relation.type)) {
-            return g(
-                    SvgTagCreator.path()
-                            .attr("d", stringPath)
-                            .attr("stroke", fillId)
-                            .attr("data-target", relation.target)
-                            .attr("data-type", relation.type)
-                    ,
-                    label(bezierPath, fillId)
-            ).attr("data-source", relation.source).attr("data-target", relation.target);
+            ContainerTag path = SvgTagCreator.path()
+                    .attr("d", stringPath)
+                    .attr("stroke", fillId);
+            if (Lifecycle.PLANNED.equals(relation.source.getLifecycle()) || Lifecycle.PLANNED.equals(relation.target.getLifecycle())) {
+                path.attr("stroke-dasharray", 10);
+                path.attr("opacity", 0.7);
+            }
+            return g(path, label(bezierPath, fillId))
+                    .attr("data-source", relation.source)
+                    .attr("data-target", relation.target);
         }
 
         List<ContainerTag> markers = new ArrayList<>();
